Issue #SC-1808: Seperate out the logic and resolved review comments
